+++ internal/app/wwctl/power/on/root.go
@@ -3,6 +3,7 @@ package on
 import (
 	"github.com/spf13/cobra"
 	"github.com/warewulf/warewulf/internal/app/wwctl/completions"
+	"github.com/warewulf/warewulf/internal/pkg/hostlist"
 )
 
 type variables struct {
@@ -16,7 +17,7 @@ func GetCommand() *cobra.Command {
 	powerCmd := &cobra.Command{
 		Use:               "on [OPTIONS] [PATTERN ...]",
 		Short:             "Power on the given node(s)",
-		Long:              "This command will power on a set of nodes specified by PATTERN.",
+		Long:              "This command will power on a set of nodes specified by PATTERN.\n" + hostlist.Docstring,
 		RunE:              CobraRunE(&vars),
 		Args:              cobra.MinimumNArgs(1),
 		ValidArgsFunction: completions.Nodes,